+++ verapdf_verifier.rb
@@ -1,14 +1,13 @@
 # frozen_string_literal: true
 
 require 'sinatra'
-require_relative 'helpers/app_helper.rb'
-
-include AppHelper
+require_relative 'lib/verapdf_verifier.rb'
+verapdf = Verapdf.new
 
 post '/payload' do
   tmp_file = Tempfile.new(['verapdf_tmp_file', '.pdf'])
   File.open(tmp_file.path, 'wb') do |f|
     f.write(params[:data])
   end
-  get_verapdf_data(tmp_file.path)
+  verapdf.file_data(tmp_file.path)
 end